build ROCM on CMAKE

+++ include/models/basellm.h
@@ -137,7 +137,7 @@ namespace fastllm {
         int head_dim = embed_dim / num_attention_heads;
         const int max_positions = 32768;
         int rotary_dim = 64;
-        const float scale_attn = sqrt(head_dim);
+        const float scale_attn = sqrtf(head_dim);
         int block_cnt = 28;
 
         std::vector<std::vector<float> > sin, cos;
